snapshot of project "lynx", label v2-8-3dev_19

+++ WWW/Library/Implementation/HTRules.c
@@ -472,16 +472,16 @@ PUBLIC int  HTSetConfiguration ARGS1(
 	else status = 0;
 	HTSetSuffix(word2,	word3,
 				encoding ? encoding : "binary",
-				status >= 1? quality : 1.0);
+				status >= 1? quality : (float) 1.0);
 
     } else if (0==strcasecomp(word1, "presentation")) {
 	if (pointer) status = sscanf(pointer, "%f%f%f%d",
 			    &quality, &secs, &secs_per_byte, &maxbytes);
 	else status = 0;
 	HTSetPresentation(word2, word3,
-		    status >= 1 ? quality		: 1.0,
-		    status >= 2 ? secs			: 0.0,
-		    status >= 3 ? secs_per_byte 	: 0.0,
+		    status >= 1 ? quality		: (float) 1.0,
+		    status >= 2 ? secs			: (float) 0.0,
+		    status >= 3 ? secs_per_byte 	: (float) 0.0,
 		    status >= 4 ? maxbytes		: 0 );
 
     } else if (0==strncasecomp(word1, "htbin", 5) ||
